requirements.txt
mock==2.0.0
pytest==3.2.1
python-slugify==1.2.4
requests==2.18.4